backend/requirements.txt
moralis
python-dotenv
requests
google-generativeai
openai
